requirements.txt
openai
python-dotenv
pandas
numpy
matplotlib
scikit-learn
statsmodels
prophet
streamlit
pytesseract
pdf2image
reportlab
weasyprint
ics
textblob
transformers
flask
pytest
